Split cmdline.py into cli_parser.py (engine) + cli_commands.py (catalog)
The big `COMMANDS` table is pure configuration and was driving most of
the churn in `cmdline.py`. Pulling it into its own module leaves
`cli_parser.py` as the parser engine plus the spec types it consumes,
and lets the two layers (what the CLI accepts vs. how arguments are
parsed) change independently.

`parse_cmdline` now takes a `ParserContext` carrying
`commands_by_name_or_alias` and `common_options`, so the parser stays
decoupled from the catalog.
